Added the background and changed the sprite black
In the creation of the sprites in piskelapp.com 0x000000 was used as black, but the gba thinks that 0x000000 is see-through. To solve this issue all the black parts are now 0x101010. You can't spot the difference while looking at it, but this small change will show the black parts of the sprites.